add: 인벤토리에 아무것도 없으면 그냥 204, total_price 구하기, receipts views fix: related_name inventory > inventories

+++ inventories/views.py
@@ -57,18 +57,21 @@ class InventoriesPurchase(APIView):
 
     def post(self, request):
         my_inventories = Inventory.objects.filter(owner=request.user)
-        receipt = Receipt.objects.create(owner=request.user)
-        for my_inventory in my_inventories:
-            receipt.past_medicine.add(
-                PastMedicine.objects.create(
-                    receipt=receipt,
-                    medicine=my_inventory.medicine,
-                    quantity=my_inventory.quantity,
-                    price_per_medicine_at_purchase=my_inventory.medicine.price,
+        if my_inventories:
+            receipt = Receipt.objects.create(owner=request.user)
+            for my_inventory in my_inventories:
+                receipt.past_medicines.add(
+                    PastMedicine.objects.create(
+                        receipt=receipt,
+                        medicine=my_inventory.medicine,
+                        quantity=my_inventory.quantity,
+                        price_per_medicine_at_purchase=my_inventory.medicine.price,
+                    )
                 )
-            )
-            my_inventory.delete()
-        return Response(status=status.HTTP_200_OK)
+                my_inventory.delete()
+            return Response(status=status.HTTP_200_OK)
+        else:
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class InventoryDetail(APIView):
